# openapi.yaml
openapi: '3.0.0'
info:
  title: Priority ERP Mock API
  version: '1.0.0'
  description: Mock Priority ERP OData-style API for workshop on building AI agents with Google Gemini SDK
servers:
  - url: http://localhost:3000
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    Customer:
      type: object
      properties:
        CUSTNAME:
          type: string
        CUSTDES:
          type: string
        PHONE:
          type: string
        EMAIL:
          type: string
        ADDRESS:
          type: string
        CITY:
          type: string
        STATE:
          type: string
        COUNTRY:
          type: string
        BALANCE:
          type: number
        CREDIT_LIMIT:
          type: number
        STATUS:
          type: string
          enum: [A, I]
    CustomerList:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
...
    BOM:
      type: object
      properties:
        BOMNAME:
          type: string
        PARTNAME:
          type: string
        BOMDES:
          type: string
        REVISION:
          type: string
        STATUS:
          type: string
        BOMITEMS_SUBFORM:
          type: array
          items:
            $ref: '#/components/schemas/BOMItem'
    BOMList:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/BOM'
    BOMItem:
      type: object
      properties:
        LINE:
          type: integer
        CHILDPART:
          type: string
        TQUANT:
          type: number
        UNIT:
          type: string
...
    Employee:
      type: object
      properties:
        EMPNUM:
          type: string
        EMPNAME:
          type: string
        DEPT:
          type: string
        POSITION:
          type: string
        HIREDATE:
          type: string
          format: date
        PHONE:
          type: string
        EMAIL:
          type: string
        STATUS:
          type: string
    EmployeeList:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
...
    Department:
      type: object
      properties:
        DEPTCODE:
          type: string
        DEPTDES:
          type: string
        MANAGER:
          type: string
        HEADCOUNT:
          type: integer
    DepartmentList:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Department'
  parameters:
    filterParam:
      name: $filter
      in: query
      description: OData filter expression
      schema:
        type: string
    orderbyParam:
      name: $orderby
      in: query
      description: OData orderby expression
      schema:
        type: string
    topParam:
      name: $top
      in: query
      description: Number of records to return
      schema:
        type: integer
    skipParam:
      name: $skip
      in: query
      description: Number of records to skip
      schema:
        type: integer
    selectParam:
      name: $select
      in: query
      description: Comma-separated list of fields to return
      schema:
        type: string
    expandParam:
      name: $expand
      in: query
      description: Subform to expand inline
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/CUSTOMERS:
    get:
      summary: List customers
      operationId: listCustomers
      tags: [Customers]
      parameters:
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/orderbyParam'
        - $ref: '#/components/parameters/topParam'
        - $ref: '#/components/parameters/skipParam'
        - $ref: '#/components/parameters/selectParam'
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a customer
      operationId: createCustomer
      tags: [Customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  "/api/CUSTOMERS('{CUSTNAME}')":
    get:
      summary: Get a customer by key
      operationId: getCustomer
      tags: [Customers]
      parameters:
        - name: CUSTNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a customer
      operationId: updateCustomer
      tags: [Customers]
      parameters:
        - name: CUSTNAME
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a customer
      operationId: deleteCustomer
      tags: [Customers]
      parameters:
        - name: CUSTNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Customer deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== ORDERS ====================
...
  /api/BOM:
    get:
      summary: List bills of materials
      operationId: listBOMs
      tags: [Manufacturing]
      parameters:
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/orderbyParam'
        - $ref: '#/components/parameters/topParam'
        - $ref: '#/components/parameters/skipParam'
        - $ref: '#/components/parameters/selectParam'
        - $ref: '#/components/parameters/expandParam'
      responses:
        '200':
          description: List of BOMs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BOMList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a bill of materials
      operationId: createBOM
      tags: [Manufacturing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BOM'
      responses:
        '201':
          description: BOM created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BOM'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  "/api/BOM('{BOMNAME}')":
    get:
      summary: Get a BOM by key
      operationId: getBOM
      tags: [Manufacturing]
      parameters:
        - name: BOMNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: BOM details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BOM'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a BOM
      operationId: updateBOM
      tags: [Manufacturing]
      parameters:
        - name: BOMNAME
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BOM'
      responses:
        '200':
          description: BOM updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BOM'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a BOM
      operationId: deleteBOM
      tags: [Manufacturing]
      parameters:
        - name: BOMNAME
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: BOM deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/EMPLOYEES:
    get:
      summary: List employees
      operationId: listEmployees
      tags: [HR]
      parameters:
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/orderbyParam'
        - $ref: '#/components/parameters/topParam'
        - $ref: '#/components/parameters/skipParam'
        - $ref: '#/components/parameters/selectParam'
      responses:
        '200':
          description: List of employees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create an employee
      operationId: createEmployee
      tags: [HR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  "/api/EMPLOYEES('{EMPNUM}')":
    get:
      summary: Get an employee by key
      operationId: getEmployee
      tags: [HR]
      parameters:
        - name: EMPNUM
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update an employee
      operationId: updateEmployee
      tags: [HR]
      parameters:
        - name: EMPNUM
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete an employee
      operationId: deleteEmployee
      tags: [HR]
      parameters:
        - name: EMPNUM
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Employee deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== TIMESHEETS ====================
...
  /api/DEPARTMENTS:
    get:
      summary: List departments
      operationId: listDepartments
      tags: [HR]
      parameters:
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/orderbyParam'
        - $ref: '#/components/parameters/topParam'
        - $ref: '#/components/parameters/skipParam'
        - $ref: '#/components/parameters/selectParam'
      responses:
        '200':
          description: List of departments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a department
      operationId: createDepartment
      tags: [HR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
      responses:
        '201':
          description: Department created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  "/api/DEPARTMENTS('{DEPTCODE}')":
    get:
      summary: Get a department by key
      operationId: getDepartment
      tags: [HR]
      parameters:
        - name: DEPTCODE
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Department details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a department
      operationId: updateDepartment
      tags: [HR]
      parameters:
        - name: DEPTCODE
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
      responses:
        '200':
          description: Department updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a department
      operationId: deleteDepartment
      tags: [HR]
      parameters:
        - name: DEPTCODE
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Department deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'